requirements.txt.txt
tkinter==8.6
tkcalendar==1.5.0
Pillow==9.2.0
numpy==1.23.2
opencv-python==4.4.0
dlib==19.24.0
requests==2.28.1